re-organize utils
